requirements.txt
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple
#requirements
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
pillow==10.4.0
numpy==1.26.4
torch==2.3.1
transformers==4.43.3
diffusers==0.30.3
accelerate==0.33.0
safetensors==0.4.5
google-generativeai==0.7.2
